apps/api/pyproject.toml
[project]
name = "oie-api"
version = "0.1.0"
description = "OIE FastAPI application serving the REST and WebSocket API"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "redis",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
